+++ src/datasources/results/ResultsDataSourceBase.ts
@@ -2,6 +2,8 @@ import { DataSourceBase } from "core/DataSourceBase";
 import { DataQueryRequest, DataFrameDTO, TestDataSourceResponse } from "@grafana/data";
 import { ResultsQuery } from "./types/types";
 import { BatchQueryConfig, QueryResponse } from "./types/QuerySteps.types";
+import { QueryBuilderOption, Workspace } from "core/types";
+import { getVariableOptions } from "core/utils";
 
 export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery> {
   baseUrl = this.instanceSettings.url + '/nitestmonitor';
@@ -13,6 +15,9 @@ export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery>
 
   private fromDateString = '${__from:date}';
   private toDateString = '${__to:date}';
+
+  readonly globalVariableOptions = (): QueryBuilderOption[] => getVariableOptions(this);
+  readonly workspacesCache = new Map<string, Workspace>([]);
 
   abstract runQuery(query: ResultsQuery, options: DataQueryRequest): Promise<DataFrameDTO>;
 
@@ -85,6 +90,19 @@ export abstract class ResultsDataSourceBase extends DataSourceBase<ResultsQuery>
       data: queryResponse,
       totalCount,
     };
+  }
+
+  async loadWorkspaces(): Promise<void> {
+    if (this.workspacesCache.size > 0) {
+      return;
+    }
+
+    const workspaces = await this.getWorkspaces()
+      .catch(error => {
+        throw new Error(error);
+      });
+
+    workspaces?.forEach(workspace => this.workspacesCache.set(workspace.id, workspace));
   }
 
   private async delay(timeout: number): Promise<void> {