+++ src/datasources/results/query-handlers/query-results/QueryResultsDataSource.ts
@@ -6,8 +6,11 @@ import { defaultResultsQuery } from "datasources/results/defaultQueries";
 
 export class QueryResultsDataSource extends ResultsDataSourceBase {
   queryResultsUrl = this.baseUrl + '/v2/query-results';
+  queryResultsValuesUrl = this.baseUrl + '/v2/query-result-values';
 
   defaultQuery = defaultResultsQuery;
+
+  readonly partNumbersCache: string[] = [];
 
   async queryResults(
     filter?: string,
@@ -32,6 +35,9 @@ export class QueryResultsDataSource extends ResultsDataSourceBase {
   }
 
   async runQuery(query: QueryResults, options: DataQueryRequest): Promise<DataFrameDTO> {
+    await this.getPartNumbers();
+    await this.loadWorkspaces();
+
     const responseData = await this.queryResults(
       this.getTimeRangeFilter(options, query.useTimeRange, query.useTimeRangeFor),
       query.orderBy,
@@ -76,6 +82,20 @@ export class QueryResultsDataSource extends ResultsDataSourceBase {
         fields: [{ name: 'Total count', values: [responseData.totalCount] }],
       };
     }
+  }
+
+  async getPartNumbers(): Promise<void> {
+    if (this.partNumbersCache.length > 0) {
+      return;
+    }
+    
+    const partNumbers = await this.post<string[]>(this.queryResultsValuesUrl, {
+      field: ResultsPropertiesOptions.PART_NUMBER,
+    }).catch(error => {
+      throw new Error(error);
+    });
+
+    partNumbers?.forEach(partNumber => this.partNumbersCache.push(partNumber));
   }
 
   shouldRunQuery(_: QueryResults): boolean {